Move services to separate module in nzbget (#146093)
